recipes/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==4.1.0
django-dotenv==1.4.2
django-environ==0.10.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
importlib-metadata==6.0.0
inflection==0.5.1
Markdown==3.4.1
packaging==23.1
Pillow==9.4.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
zipp==3.15.0
